Add: jpa repositories (Game, MoveBook)

+++ src/main/java/com/project/chess/data/repository/UserRepository.java
@@ -3,8 +3,6 @@ package com.project.chess.data.repository;
 import com.project.chess.data.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<User, Integer>  {
-
+public interface UserRepository extends JpaRepository<User, Long>  {
    User getByUsername(String username);
-
 }
